[mono] Set default $(TargetFrameworkFallbackSearchPaths) for osx
- This sets up the $(TargetFrameworkFallbackSearchPaths) to point to
  `/Library/Frameworks/Mono.framework/External/xbuild-frameworks` on
  mono/osx. This location is used by Xamarin to install frameworks.
  The .net frameworks are available in usual default location.

  So, effectively, target frameworks are looked up, in order:

  1. $(TargetFrameworkRootPath) or the default location, if
	 $(TargetFrameworkRootPath) is ''

  2. If not found in (1), then try to find it in the fallback search
	 paths specified in $(TargetFrameworkFallbackSearchPaths)

Note that we always add the `External` as the last fallback path on
`$(TargetFrameworkFallbackSearchPaths)`.

Rationale:

When running on Mono, target frameworks are searched for in
`$mono_prefix/lib/mono/xbuild-frameworks`, which becomes
`/Library/Frameworks/Mono.framework/Versions/Current/lib/mono/xbuild-frameworks`.

This works, but has a problem when upgrading the Mono.framework package:
a *new* `/Library/Frameworks/Mono.framework/Versions/@VERSION@` directory
is created, so anything installed into the previous directory will be
"lost". The file is still there, but xbuild will no longer find it.

To address this upgrade scenario, Mono's xbuild also checks a fallback path:

	`/Library/Frameworks/Mono.framework/External/xbuild-frameworks`

This location is not within a versioned framework directory, and thus files
installed into it will survive Mono.framework package upgrades.

This also allows overriding the root path to test in-tree builds of custom target
frameworks. Simply changing `$(TargetFrameworkRootPath)` would mean that
the in-tree builds of the custom frameworks would work but the default
ones would not be found!