Allow overriding fallback paths for MSBuildExtensionsPath via a property (#59)
* Add support for fallback paths to be specified via property references

Currently, the list of fallback paths in app.config can have a property
reference, like:

```
 <searchPaths os="osx">
	<property name="MSBuildExtensionsPath" value="$(FallbackPath);/x/y/z" />
```

.. but it is assumed that such a property will expand to a single path.

This commit adds support for such properties to expand to multiple ';'
separated list of paths. The rules for expansion do not change, they are still
expanded at the time of project import.

With this, given `$(FallbackPath)="/a/b/c;/q/w/e"` and the list of fallback paths
would become `"/a/b/c;/q/w/e;/x/yz"`.

* [osx] Add property allowing users to override fallback paths for ..

.. MSBuildExtensionsPath.

For OSX, this essentially sets the fallback paths to:

`<property name="MSBuildExtensionsPath" value="$(MSBuildExtensionsPathFallbackPathsOverride);/Library/Frameworks/Mono.framework/External/xbuild/" />`

.. whichs allows the default fallback paths to be overridden by setting
a property - `$(MSBuildExtensionsPathFallbackPathsOverride)`.

One of the useful use cases for this is to be able to run tests with
in-tree builds of assemblies/targets meant for
`$(MSBuildExtensionsPath)/`. Simply changing `$(MSBuildExtensionsPath)` would
mean that the in-tree builds of the custom frameworks would work but the
default ones would not be found!

(cherry picked from commit c4e70a612bbe5454e7ddc6189b9ea8dce01d17c8)